online_inference/requirements.txt
anyio==3.6.1
asgiref==3.5.2
atomicwrites==1.4.0
attrs==21.4.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
fastapi==0.78.0
h11==0.13.0
idna==3.3
imbalanced-learn==0.9.0
imblearn==0.0
iniconfig==1.1.1
joblib==1.1.0
numpy==1.22.4
packaging==21.3
pandas==1.4.2
pluggy==1.0.0
py==1.11.0
pydantic==1.9.1
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.1
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
threadpoolctl==3.1.0
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
